requirements.txt
backcall==0.2.0
coverage==3.6
decorator==5.1.0
Django==2.2.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==19.9.0
html-testRunner==1.2.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.3
MarkupSafe==2.0.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg-resources==0.0.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.10.0
python-decouple==3.5
pytz==2021.1
qrcode==7.3.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
urllib3==1.26.8
wcwidth==0.2.5
